+++ core/src/main/java/nl/nn/adapterframework/processors/PipeProcessorBase.java
@@ -15,6 +15,7 @@
 */
 package nl.nn.adapterframework.processors;
 
+import nl.nn.adapterframework.util.AppConstants;
 import nl.nn.adapterframework.util.LogUtil;
 
 import org.apache.log4j.Logger;
@@ -27,6 +28,9 @@ import org.apache.log4j.Logger;
  */
 public abstract class PipeProcessorBase implements PipeProcessor {
 	protected Logger log = LogUtil.getLogger(this);
+	protected Logger secLog = LogUtil.getLogger("SEC");
+
+	protected boolean secLogEnabled = AppConstants.getInstance().getBoolean("sec.log.enabled", false);
 
 	protected PipeProcessor pipeProcessor;